src/pool-detection/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
onnxruntime>=1.17.0
numpy>=1.26
Pillow>=10.0
arq>=0.26
pytest>=8.0
ruff>=0.4
